+++ CHANGELOG.md
@@ -12,9 +12,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - replace Cesium glb/gltf conversion with much simpler https://github.com/najadojo/gltf-import-export
 - Right-click `Import Materials`. Reorders material list and fills some gaps.
 - Right-click `Build Anno Mod` on `annomod.json`
-- show file path in outline instead of only file names (@Taubenangriff)
 - support glTF binary in `gltf` converter
-- asset GUID hover / auto-conversion
+
+### Added
+
+- Setting to control folder depth shown in .cfg outline
+- Asset name on hover over GUIDs in XPath strings and in tags `Ingredient`, `Product`, `ItemLink` or `GUID`.
+  Supported assets are limited to products, items, buildings, pools (to maintain speed and usability).
+- Asset name to GUID conversion when in tags (same as hover).
+  Supported assets are limited to products and items.
 
 ## [0.0.11]